refactor(backend): extract entity-revision engine into farm/history
Move the revision registry, snapshot/diff helpers, record_entity_revision
and whole-project restore out of farm/views.py into a dedicated
farm/history package (records.py + restore.py). Code moves verbatim;
farm/views.py now imports the seam from farm.history. No URL, payload
or behavior changes.
