+++ backend/farm/history/__init__.py
@@ -1,0 +1,27 @@
+"""Entity-revision engine for per-entity history and point-in-time restore.
+
+Public seam used by the API layer (views/mixins): recording revisions,
+serializing snapshots, and restoring whole-project state.
+"""
+
+from .records import (
+    _build_entity_revision_changes,
+    _current_actor_label,
+    _diff_changed_fields,
+    _entity_display_name,
+    _entity_type_for,
+    _serialize_instance,
+    record_entity_revision,
+)
+from .restore import _restore_project_state_at
+
+__all__ = [
+    '_build_entity_revision_changes',
+    '_current_actor_label',
+    '_diff_changed_fields',
+    '_entity_display_name',
+    '_entity_type_for',
+    '_restore_project_state_at',
+    '_serialize_instance',
+    'record_entity_revision',
+]
